third commit

+++ mainwindow.cpp
@@ -12,8 +12,8 @@ MainWindow::MainWindow(QWidget *parent)
 {
     ui->setupUi(this);
 
-    m_listWidgetCollection = {
-            new ListDecorator(ui->listWidget1, "/", true), new ListDecorator(ui->listWidget2, "/", false)};
+    m_activeList = ui->listWidget1;
+    m_passiveList = ui->listWidget2;
 
     init();
 }
@@ -23,24 +23,22 @@ MainWindow::~MainWindow()
     delete ui;
 }
 
-
-//void MainWindow::init() {
-
-//}
-
-//ListDecorator *MainWindow::getActiveList() {
-//    return m_listWidgetCollection[0]->isActive ? m_listWidgetCollection[0] : m_listWidgetCollection[1];
-//}
+void MainWindow::showException(QString message) {
+    QMessageBox::warning(this, tr("Error"), message);
+}
 
 
 void MainWindow::init() {
-
-    generateFilePanel1(m_activeList, "/");
-    generateFilePanel2(m_passiveList, "/");
-
     ui->menuFile->addAction(ui->actionRemove);
     ui->menuFile->addAction(ui->actionCopy);
     ui->menuFile->addAction(ui->actionMove);
+    ui->menuFile->addAction(ui->actionRename);
+    ui->menuFile->addAction(ui->actionNew_Directory);
+//    ui->menuFile->addAction(ui->actionNew_File);
+    ui->menuFile->addAction(ui->actionSearch);
+
+    generateFilePanel1(m_activeList, "/home/papka");
+    generateFilePanel2(m_passiveList, "/home/papka/Downloads");
 
     // enter and double click
     connect(m_activeList, &QListWidget::itemActivated, this, &MainWindow::processItem1);
@@ -49,6 +47,10 @@ void MainWindow::init() {
     connect(ui->actionCopy, &QAction::triggered, this, &MainWindow::copyItem);
     connect(ui->actionMove, &QAction::triggered, this, &MainWindow::moveItem);
     connect(ui->actionRemove, &QAction::triggered, this, &MainWindow::removeItem);
+    connect(ui->actionRename, &QAction::triggered, this, &MainWindow::renameItem);
+    connect(ui->actionNew_Directory, &QAction::triggered, this, &MainWindow::makeDir);
+//    connect(ui->actionNew_File, &QAction::triggered, this, &MainWindow::makeFile);
+    connect(ui->actionSearch, &QAction::triggered, this, &MainWindow::searchItem);
 
 }
 
@@ -83,10 +85,21 @@ void MainWindow::copyItem() {
     auto itemIndex = item->listWidget()->row(item) + 1;
 
     QString newPath = m_currDirVector[1].path() + "/" + files[itemIndex].fileName();
-    if (QFile::exists(newPath))
-    {
-        QFile::remove(newPath);
+    if (files[itemIndex].isFile()) {
+        if (QFile::exists(newPath))
+        {
+            QFile::remove(newPath);
+        }
+
+        QFile::copy(files[itemIndex].absoluteFilePath(), newPath);
     }
+
+    else if(files[itemIndex].isDir()) {
+        if (!copyRecursively(files[itemIndex].absoluteFilePath(), newPath)) {
+            showException("Cannot copy this directory");
+        }
+    }
+
 
     QFile::copy(files[itemIndex].absoluteFilePath(), newPath);
 
@@ -130,13 +143,105 @@ void MainWindow::removeItem() {
 
     auto files = m_filesVector[0];
     auto itemIndex = item->listWidget()->row(item) + 1;
-    if (QFile::exists(files[itemIndex].absoluteFilePath()))
+    if (files[itemIndex].isFile() && QFile::exists(files[itemIndex].absoluteFilePath()))
     {
-        QFile::remove(files[itemIndex].absoluteFilePath());
+        if (!QFile::remove(files[itemIndex].absoluteFilePath())){
+            showException("Cannot remove this file");
+        }
+    }
+
+    else if (files[itemIndex].isDir() && QFile::exists(files[itemIndex].absoluteFilePath()))
+    {
+        if (!QDir(files[itemIndex].absoluteFilePath()).removeRecursively()){
+            showException("Cannot remove this dir");
+        }
     }
 
     generateFilePanel1(m_activeList, m_currDirVector[0].path());
     generateFilePanel2(m_passiveList, m_currDirVector[1].path());
+}
+
+void MainWindow::makeDir() {
+
+    QString newDirName = QInputDialog::getText(this,"Title","text");
+
+    QListWidgetItem *item = m_activeList->currentItem();
+
+    auto files = m_filesVector[0];
+    auto itemIndex = item->listWidget()->row(item) + 1;
+    if (QDir(files[itemIndex].absolutePath()).exists())
+    {
+        if (!QDir().mkdir(files[itemIndex].absolutePath() + "/" + newDirName)){
+            showException("Cannot create directory");        }
+    }
+
+    generateFilePanel1(m_activeList, m_currDirVector[0].path());
+    generateFilePanel2(m_passiveList, m_currDirVector[1].path());
+
+}
+
+void MainWindow::makeFile() {
+
+    QString newFileName = QInputDialog::getText(this,"Title","text");
+
+    QListWidgetItem *item = m_activeList->currentItem();
+
+    auto files = m_filesVector[0];
+    auto itemIndex = item->listWidget()->row(item) + 1;
+    if (QDir(files[itemIndex].absolutePath()).exists())
+    {
+        QFile filepath;
+        filepath.setFileName(files[itemIndex].absolutePath() + "/" + newFileName);
+        if (!filepath.open(QIODevice::WriteOnly|QIODevice::Text)){
+            showException("Cannot create file");        }
+    }
+
+    generateFilePanel1(m_activeList, m_currDirVector[0].path());
+    generateFilePanel2(m_passiveList, m_currDirVector[1].path());
+
+}
+
+void MainWindow::renameItem() {
+
+    QString newFileName = QInputDialog::getText(this,"Title","text");
+
+    QListWidgetItem *item = m_activeList->currentItem();
+
+    auto files = m_filesVector[0];
+    auto itemIndex = item->listWidget()->row(item) + 1;
+    if (QFile::exists(files[itemIndex].absoluteFilePath()))
+    {
+        if (!QFile::rename(files[itemIndex].absoluteFilePath(), files[itemIndex].absolutePath() + "/" + newFileName)){
+            showException("Cannot rename this file");
+        }
+    }
+
+    generateFilePanel1(m_activeList, m_currDirVector[0].path());
+    generateFilePanel2(m_passiveList, m_currDirVector[1].path());
+
+}
+
+void MainWindow::searchItem() {
+
+//    QString searchedFileName = QInputDialog::getText(this,"Title","text");
+
+//    QVector<QString> *resultPath = {};
+//    searchRecursively("", searchedFileName, resultPath);
+
+//    if (resultPath->size() > 0) {
+
+//        QMessageBox::warning(this, tr("Error"), resultPath->at(0));
+//    }
+//    else {
+
+//        QMessageBox::warning(this, tr("Error"), tr("File did not exist"));
+//    }
+
+    QMessageBox::warning(this, tr("Error"), tr("Manager does not support searching files"));
+
+//    generateFilePanel1(m_activeList, m_currDirVector[0].path());
+//    generateFilePanel2(m_passiveList, m_currDirVector[1].path());
+
 }
 
 void MainWindow::generateFilePanel1(QListWidget *listPanel, QString pathString) {
@@ -191,5 +296,38 @@ void MainWindow::generateFilePanel2(QListWidget *listPanel, QString pathString) 
     ui->label2->setText(pathString);
 }
 
+////////////////////////////////////////////////
+bool MainWindow::copyRecursively(QString srcFilePath, QString tgtFilePath)
+{
+    QFileInfo srcFileInfo(srcFilePath);
+    if (srcFileInfo.isDir()) {
+        QDir targetDir(tgtFilePath);
+        targetDir.cdUp();
+        if (!targetDir.mkdir(QFileInfo(tgtFilePath).fileName()))
+            return false;
+        QDir sourceDir(srcFilePath);
+        QStringList fileNames = sourceDir.entryList(QDir::Files | QDir::Dirs | QDir::NoDotAndDotDot | QDir::Hidden | QDir::System);
+        foreach (const QString &fileName, fileNames) {
+            const QString newSrcFilePath
+                    = srcFilePath + QLatin1Char('/') + fileName;
+            const QString newTgtFilePath
+                    = tgtFilePath + QLatin1Char('/') + fileName;
+            if (!MainWindow::copyRecursively(newSrcFilePath, newTgtFilePath))
+                return false;
+        }
+    } else {
+        if (!QFile::copy(srcFilePath, tgtFilePath))
+            return false;
+    }
+    return true;
+}
 
-
+void MainWindow::searchRecursively(const QString &path, const QString &pattern, QVector<QString> *result)
+{
+    QDir currentDir(path);
+    const QString prefix = path + "/";
+    foreach (const QString &match, currentDir.entryList(QStringList(pattern), QDir::Files | QDir::NoSymLinks))
+        result->append(prefix + match);
+    foreach (const QString &dir, currentDir.entryList(QDir::Dirs | QDir::NoSymLinks | QDir::NoDotAndDotDot))
+        MainWindow::searchRecursively(prefix + dir, pattern, result);
+}
